+++ Craftsman/Builders/Tests/Utilities/UnitTestUtilsBuilder.cs
@@ -37,22 +37,26 @@ public static class UnitTestUtilsBuilder
 using System.Reflection;
 using Mapster;
 using MapsterMapper;
+using Services;
 
 public class UnitTestUtils
 {{
     public static Mapper GetApiMapper()
     {{
-        var apiAssembly = GetAssemblyByName(""{_scaffoldingDirectoryStore.ProjectBaseName}"");
+        var apiAssembly = GetApiAssembly();
         var typeAdapterConfig = TypeAdapterConfig.GlobalSettings.Clone();
         typeAdapterConfig.Scan(apiAssembly);
         var mapper = new Mapper(typeAdapterConfig);
         return mapper;
     }}
 
-    private static Assembly GetAssemblyByName(string name)
+    public static Assembly GetApiAssembly()
     {{
-        return AppDomain.CurrentDomain.GetAssemblies().
-            SingleOrDefault(assembly => assembly.GetName().Name == name);
+        // need to load something from the api for it to be in the loaded assemblies
+        _ = new DateTimeProvider();
+        return AppDomain.CurrentDomain
+            .GetAssemblies()
+            .FirstOrDefault(a => a.GetName().Name == ""{_scaffoldingDirectoryStore.ProjectBaseName}"");
     }}
 }}
 ";